Move defaults to constants file
